Stable Version 2 before template designer

+++ app.py
@@ -1,5 +1,6 @@
-from flask import Flask, render_template, request, send_file
+from flask import Flask, render_template, request, send_file, jsonify
 import os
+from flask import send_from_directory
 import subprocess
 import shutil
 
@@ -31,20 +32,16 @@ def upload():
     certificate = request.files["certificate"]
     excel = request.files["excel"]
 
-    # Get settings from form
-    fontsize = request.form["fontsize"]
-    x = request.form["x"]
-    y = request.form["y"]
-
-    # Save settings
-    with open("settings.txt", "w") as f:
-        f.write(f"{fontsize}\n")
-        f.write(f"{x}\n")
-        f.write(f"{y}\n")
-
     # Save uploaded files
     certificate.save(os.path.join(UPLOAD_FOLDER, "certificate.png"))
     excel.save(os.path.join(UPLOAD_FOLDER, "Students.xlsx"))
+
+    # If settings.txt does not exist, create default values
+    if not os.path.exists("settings.txt"):
+        with open("settings.txt", "w") as f:
+            f.write("45\n")
+            f.write("1030\n")
+            f.write("535\n")
 
     # Run certificate generation
     subprocess.run(["python", "generate.py"])
@@ -63,9 +60,34 @@ def download():
         as_attachment=True
     )
 
+@app.route("/editor")
+def editor():
+
+    return render_template("editor.html")
+
+@app.route("/save_position", methods=["POST"])
+def save_position():
+
+    data = request.get_json()
+
+    x = data["x"]
+    y = data["y"]
+
+    # Save coordinates
+    with open("settings.txt", "w") as f:
+        f.write("45\n")
+        f.write(str(x) + "\n")
+        f.write(str(y) + "\n")
+
+    return "✅ Layout Saved Successfully!"
+
 @app.route("/test")
 def test():
     return "TEST ROUTE WORKING"
+
+@app.route("/uploads/<filename>")
+def uploaded_file(filename):
+    return send_from_directory("uploads", filename)
 
 # ----------------------------
 # Run Application
